style: improve code formatting and spacing in Box class

+++ src/com/google/inheritance/Box.java
@@ -1,38 +1,38 @@
 package com.google.inheritance;
 
-public class Box{
+public class Box {
     double l;
     double w;
     double h;
 
-    public Box(){
+    public Box() {
         this.l = -1;
         this.w = -1;
         this.h = -1;
     }
 
     // Cube
-    Box(double side){
+    Box(double side) {
         this.l = side;
-        this.w = side;  
+        this.w = side;
         this.h = side;
     }
 
     // rectangle
-    Box(int len, int width, int height){
+    Box(double len, double width, double height) {
         this.l = len;
         this.w = width;
         this.h = height;
     }
 
     // copy constructor
-    Box(Box old){
+    Box(Box old) {
         this.l = old.l;
         this.w = old.w;
         this.h = old.h;
     }
 
-    public void information(){
+    public void information() {
         System.out.println("Running the box...");
     }
 }
